+++ src/app/sf1/complex-search/complex-search.component.ts
@@ -45,10 +45,10 @@ export class ComplexSearchComponent implements OnInit {
   }
 
   doSearch() {
-    const k = this.exp.getKeyWords();
-    console.log(k);
+    const k = this.exp.getKeyWords()
+    console.log(k)
 
-    const exp = this.exp.getValue()
+    const exp = this.exp.Encode()
 
     // const dbs = this.exp.getDBValue()
     const dbs = 'FMZL,FMSQ,SYXX,WGZL'
@@ -63,4 +63,5 @@ export class ComplexSearchComponent implements OnInit {
 
     this.router.navigate(['/sf1/list'], { queryParams: { exp, dbs, order, option, from, to, displayCols } });
   }
+
 }