+++ src/app/sf1/sf1-search.service.ts
@@ -1,5 +1,6 @@
 import * as moment from 'moment'
-import { forEach } from '@angular/router/src/utils/collection';
+import { forEach } from '@angular/router/src/utils/collection'
+import * as crypt from 'crypt'
 
 class SF1SearchConditionItem {
   op: string
@@ -459,6 +460,195 @@ export class SF1SearchExp {
     })
 
     return this.getValueByGroup(g);
+  }
+
+  Encode(){
+    const j=JSON.stringify(this)
+    const b = [];
+    
+    for (let i = 0; i < j.length; ++i)
+    {
+        let c= j.charCodeAt(i)
+
+        if (c<128){
+          b.push(c)
+        }
+        else if(c<2048){
+          b.push((c >> 6) | 192)
+          b.push((c & 63) | 128)
+        }
+        else{
+          b.push((c >> 12) | 224)
+          b.push(((c >> 6) & 63) | 128)
+          b.push((c & 63) | 128)
+        }
+    }
+    const v =crypt.bytesToBase64(b)
+
+    return v
+  }
+
+  Decode(v:string){
+    let b=crypt.base64ToBytes(v)
+    let j=''
+    let i=0
+    while (i < b.length) {
+      let c = b[i]
+      if (c < 128) {
+        j += String.fromCharCode(c)
+        i++
+      }
+      else if ((c > 191) && (c < 224)) {
+        let c2 = b[i + 1];
+        j += String.fromCharCode(((c & 31) << 6) | (c2 & 63))
+        i += 2
+      }
+      else {
+        let c2 = b[i + 1]
+        let c3 = b[i + 2]
+        j += String.fromCharCode(((c & 15) << 12) | ((c2 & 63) << 6) | (c3 & 63))
+        i += 3
+      }
+    }
+    
+    this.setFromJson(j)
+  }
+
+
+  private setGroup(j:Array<any>){
+    let g=new Array<SF1SearchCondition>()
+
+    j.forEach((item)=>{
+      let i=new SF1SearchCondition(item.id,item.name,item.title)
+      i.items=item.items
+      g.push(i)
+    })
+
+    return g
+  }
+
+  private setFromJson(json:string)
+  {
+    let j=JSON.parse(json)
+
+    this.db_group=j.db_group  
+    this.lastKeyWord=j.lastKeyWord
+
+    this.key_group=this.setGroup(j.key_group)
+    this.code_group=this.setGroup(j.code_group)
+    this.type_group=this.setGroup(j.type_group)
+    this.name_group=this.setGroup(j.name_group)
+    this.date_group=this.setGroup(j.date_group)
+  }
+
+  private getItemDisplay(item:SF1SearchCondition){
+    let k=item.title
+    let op=''
+    let v=''
+    item.items.forEach((vi)=>{
+      if (vi.value){
+        if (!op) {
+          op=vi.op
+        }
+        else {
+          v+=' '+vi.op+' '
+        }
+        v+=vi.value
+      }
+    })
+
+    if (v) {
+      return op+' '+k+'=('+v+')'
+    }
+    else{
+      return ''
+    }
+  }
+
+  private getDateItemDisplay(item:SF1SearchCondition){
+    let k=item.title
+    let v=''
+    item.items.forEach((vi)=>{
+      let f=vi.from
+      let t=vi.to
+      let vs=''
+      
+      switch (vi.mode){
+          case '0':
+            if (f && t){
+              vs='(from '+moment(f).format('YYYYMMDD') + ' to ' + moment(t).format('YYYYMMDD')+')'
+            }
+            break
+          case '1':
+            if (t){
+              vs='(to ' + moment(t).format('YYYYMMDD')+')'
+            }
+            break
+          case '2':
+            if (f){
+              vs='(from '+moment(f).format('YYYYMMDD')+')'
+            }
+            break
+          case '3':
+            if (f){
+              vs='('+moment(f).format('YYYYMMDD')+')'
+            }
+            break
+      }
+
+      if (vs){
+          v+=(' '+vi.op+' '+k+'='+vs)
+      }
+    })
+
+    return v
+  }
+
+  private getGroupDisplay(g:Array<SF1SearchCondition>){
+    let r=new Array<string>()
+
+    g.forEach((c)=>{
+      let v=this.getItemDisplay(c)
+      if (v){
+        r.push(v)
+      }
+    })
+
+    return r
+  }
+
+  private getDateDisplay(g:Array<SF1SearchCondition>){
+    let r=new Array<string>()
+
+    g.forEach((c)=>{
+      let v=this.getDateItemDisplay(c)
+      if (v){
+        r.push(v)
+      }
+    })
+
+    return r
+  }
+
+  getDisplay(){
+    let r=new Array<string>()
+
+    r=r.concat(this.getGroupDisplay(this.key_group))
+    r=r.concat(this.getGroupDisplay(this.code_group))
+    r=r.concat(this.getGroupDisplay(this.type_group))
+    r=r.concat(this.getGroupDisplay(this.name_group))
+    r=r.concat(this.getDateDisplay(this.date_group))
+
+    if (r.length>0){
+      if (r[0].startsWith('AND') || r[0].startsWith('NOT')){
+        r[0]=r[0].substr(4,r[0].length-4)
+      }
+      else if (r[0].startsWith('OR')){
+        r[0]=r[0].substr(3,r[0].length-3)
+      }
+    }
+
+    return r
   }
 
 }